:bug: nice hmr

+++ apps/models/src/index.js
@@ -82,7 +82,6 @@ export default class Database {
     if (Constants.IS_BROWSER) {
       RxDB.QueryChangeDetector.enable()
     }
-    console.log('Create RxDB with adapter', this.databaseConfig.adapterName)
     RxDB.plugin(this.databaseConfig.adapter)
     RxDB.plugin(pHTTP)
     PouchDB.plugin(this.databaseConfig.adapter)
@@ -104,14 +103,16 @@ export default class Database {
     this.connected = true
   }
 
-  dispose = async () => {
+  async dispose() {
     for (const name of Object.keys(this.models)) {
       const model = this.models[name]
       if (model) {
         await model.dispose()
       }
     }
-    await this.database.destroy()
+    if (this.database) {
+      await this.database.destroy()
+    }
   }
 
   attachModels = async () => {
